refactor: moving things around
