+++ parser/parser_test.go
@@ -8,30 +8,30 @@ import (
 
 var (
 	qualityParserTests = map[string]parser.Result{
-		"Some Title":                 {},
-		"Some Title HDTV":            {Source: parser.HDTV},
-		"Some Title PDTV":            {Source: parser.PDTV},
-		"Some Title SDTV":            {Source: parser.SDTV},
-		"Some Title TVRip":           {Source: parser.TV},
-		"Some Title BD":              {Source: parser.BluRay},
-		"Some Title BR-Rip":          {Source: parser.BluRay},
-		"Some Title Blu-Ray":         {Source: parser.BluRay},
-		"Some Title DVD":             {Source: parser.DVD},
-		"Some Title.avi 720p":        {Resolution: parser.R720},
-		"Some Title.avi 720p webdl":  {Resolution: parser.R720, Source: parser.WEB},
-		"1080p web dl of Some Title": {Resolution: parser.R1080, Source: parser.WEB},
-		"1080p.web-dl.of.A.Movie":    {Resolution: parser.R1080, Source: parser.WEB},
-		"Some Title repack":          {Proper: true},
-		"Some.WEB.DL-HUNDUB.1080P":   {Resolution: parser.R1080, Source: parser.WEB, Language: parser.Hungarian},
-		"Some.Title.2012.Remux":      {Year: 2012, Remux: true, Title: "some title"},
-		"Some.Title.S01E02":          {Season: 1, Episode: 2},
+		"/some/dir/Some Title":                 {},
+		"/some/dir/Some Title HDTV":            {Source: parser.HDTV},
+		"/some/dir/Some Title PDTV":            {Source: parser.PDTV},
+		"/some/dir/Some Title SDTV":            {Source: parser.SDTV},
+		"/some/dir/Some Title TVRip":           {Source: parser.TV},
+		"/some/dir/Some Title BD":              {Source: parser.BluRay},
+		"/some/dir/Some Title BR-Rip":          {Source: parser.BluRay},
+		"/some/dir/Some Title Blu-Ray":         {Source: parser.BluRay},
+		"/some/dir/Some Title DVD":             {Source: parser.DVD},
+		"/some/dir/Some Title.avi 720p":        {Resolution: parser.R720},
+		"/some/dir/Some Title.avi 720p webdl":  {Resolution: parser.R720, Source: parser.WEB},
+		"/some/dir/1080p web dl of Some Title": {Resolution: parser.R1080, Source: parser.WEB},
+		"/some/dir/1080p.web-dl.of.A.Movie":    {Resolution: parser.R1080, Source: parser.WEB},
+		"/some/dir/Some Title repack":          {Proper: true},
+		"/some/dir/Some.WEB.DL-HUNDUB.1080P":   {Resolution: parser.R1080, Source: parser.WEB, Language: parser.Hungarian},
+		"/some/dir/Some.Title.2012.Remux":      {Year: 2012, Remux: true, Title: "some title"},
+		"/some/dir/Some.Title.S01E02":          {Season: 1, Episode: 2},
 	}
 )
 
 func TestParse(t *testing.T) {
 	for title, expected := range qualityParserTests {
 		t.Logf("Testing title: %s", title)
-		got := parser.Parse(title, "", parser.Result{})
+		got := parser.Parse(title, "/dev/null", parser.Result{})
 		compareResult(t, &expected, got)
 	}
 }
@@ -80,7 +80,7 @@ func TestOverride(t *testing.T) {
 		Proper:       true,
 		DualLanguage: true,
 	}
-	result := parser.Parse("1080p web dl of Some Title", "", overrides)
+	result := parser.Parse("1080p web dl of Some Title", "/dev/null", overrides)
 	if overrides != *result {
 		t.Errorf("expected overrides to have an effect")
 	}
